+++ src/Popup.jsx
@@ -5,12 +5,70 @@
  */
 
 var React = require('react');
-var CSSTransitionGroup = require('rc-css-transition-group');
+var anim = require('css-animation');
 var utils = require('./utils');
+var assign = require('object-assign');
+var domAlign = require('dom-align');
 
 class Popup extends React.Component {
   getRootNode() {
     return React.findDOMNode(this.refs.popup);
+  }
+
+
+  alignRootNode() {
+    var props = this.props;
+    if (props.visible) {
+      var wrapDomNode = React.findDOMNode(props.wrap);
+      var popupDomNode = this.getRootNode();
+      var placement = props.placement;
+      var points;
+      if (placement && placement.points) {
+        var align = domAlign(popupDomNode, wrapDomNode, placement);
+        popupDomNode.className = utils.getToolTipClassByPlacement(props.prefixCls, align);
+      } else {
+        points = ['cr', 'cl'];
+        if (placement === 'right') {
+          points = ['cl', 'cr'];
+        } else if (placement === 'top') {
+          points = ['bc', 'tc'];
+        } else if (placement === 'bottom') {
+          points = ['tc', 'bc'];
+        }
+        domAlign(popupDomNode, wrapDomNode, {
+          points: points
+        });
+      }
+    }
+  }
+
+  animateRootNode(prevProps) {
+    var props = this.props;
+    var transitionName = props.transitionName;
+    if (!transitionName && props.animation) {
+      transitionName = `${props.prefixCls}-${props.animation}`;
+    }
+    if (transitionName) {
+      var domNode = this.getRootNode();
+      if (props.visible && !prevProps.visible) {
+        anim(domNode, `${transitionName}-enter`);
+      } else if (!props.visible && prevProps.visible) {
+        domNode.style.display = 'block';
+        anim(domNode, `${transitionName}-leave`, ()=> {
+          domNode.style.display = 'none';
+        });
+      }
+    }
+  }
+
+  componentDidUpdate(prevProps) {
+    prevProps = prevProps || {};
+    this.alignRootNode();
+    this.animateRootNode(prevProps);
+  }
+
+  componentDidMount() {
+    this.componentDidUpdate();
   }
 
   render() {
@@ -19,22 +77,23 @@ class Popup extends React.Component {
     if (props.className) {
       className += ' ' + props.className;
     }
+    var style = this.style;
+    if (!props.visible) {
+      style = assign({}, style, {
+        display: 'none'
+      });
+    }
     var arrowClassName = `${props.prefixCls}-arrow`;
     var innerClassname = `${props.prefixCls}-inner`;
-    var content = props.visible ? [<div className={className}
+    return <div className={className}
       key="popup"
       ref="popup"
-      style={this.style}>
+      style={style}>
       <div className={arrowClassName}></div>
       <div className={innerClassname}>
     {props.children}
       </div>
-    </div>] : [];
-    if (props.transitionName) {
-      return <CSSTransitionGroup transitionName={props.transitionName}>{content}</CSSTransitionGroup>;
-    } else {
-      return content[0] || null;
-    }
+    </div>;
   }
 }