+++ HISTORY.md
@@ -1,6 +1,12 @@
 # History
 ----
 
+### 2.0.0 / 2015-06-08
+
+- support click document to hide
+- support animation props
+- support renderPopupToBody props
+
 ## 1.1.1 / 2015-05-14
 
 add defaultVisible and onVisibleChange